+++ CHANGELOG.rst
@@ -1,6 +1,8 @@
 new version (on deck)
 =====================
 - [ENH] Added function ``sort_timestamps_monotonically`` to timeseries functions @UGuntupalli
+- [ENH] Added the complete function for converting implicit missing values 
+  to explicit ones. @samukweku
 - [ENH] Further simplification of expand_grid. @samukweku
 
 v0.20.9